pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "redrat-ir-controller"
version = "0.1.0"
description = "RedRat3 IR transceiver controller with voice command support"
requires-python = ">=3.11"
dependencies = [
    "pyusb>=1.2.1",
    "PyYAML>=6.0",
    "flask>=3.0",
    "sounddevice>=0.5.0",
    "numpy>=1.24,<2",
    "vosk>=0.3.45",
    "openwakeword>=0.6.0",
    "rapidfuzz>=3.9",
]

[project.scripts]
redrat-controller = "main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["redrat*", "macros*", "voice*", "api*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
